I-#78: Remove unnessary type infering, renameing file moving file from one to another.
