+++ core/application/userWorkflows/UserService.ts
@@ -1,7 +1,7 @@
 import { GENERIC_CODES } from '@commons/libs/constants/GenericCodes'
 import UserOperationError from './UserOperationError'
 import { UserDTO } from '@commons/dto/UserDTO'
-import { generateUniqueID } from '../utils/ksuidGenerator'
+import { generateUniqueID } from '../utils/idGenerator'
 import { GenericMessage } from '@commons/dto/MessageDTO'
 import { getEmailVerificationMessage, getPasswordResetVerificationMessage } from './userMessages'
 import Repository from '@application/technicalImpl/policies/repositories/Repository'
